:recycle: [refactor] 홈 화면의 비디오 프레임 위젯 분리 (HH-176)
